+++ PrivateWin10/Controls/ControlList.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using MiscHelpers;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,6 +7,7 @@ using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
+using System.Windows.Threading;
 
 namespace PrivateWin10
 {